structure corrected
