+++ .gitignore
@@ -4,7 +4,6 @@
 
 # IntelliJ project files
 .idea
-*.iml
 out
 gen
 # Logs
@@ -19,5 +18,3 @@ node_modules
 .npm
 
 # Optional REPL history
-.DS_Store
-.vscode
