+++ app_server/controllers/userController.js
@@ -1,42 +1,10 @@
-const fs = require("fs");
 const path = require("path");
-
+const readBase = require(path.join(__dirname, "../models", "readBase"));
 const usersPath = path.join(__dirname, "../data/users.json");
-const cardsPath = path.join(__dirname, "../data/cards.json");
-
-// function for reading files
-function readBase(url) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(url, { encoding: "utf-8" }, (err, data) => {
-      if (err) {
-        reject(err); // in the case of error, control flow goes to the catch block with the error occured.
-      } else {
-        resolve(data); // in the case of success, control flow goes to the then block with the content of the file.
-      }
-    });
-  });
-}
-
-exports.getCards = (req, res) => {
-  readBase(cardsPath)
-    .then((data) => {
-      const cardsSchedule = JSON.parse(data);
-      res.status(200).json({
-        status: "success",
-        results: cardsSchedule.length,
-        data: {
-          cardsSchedule,
-        },
-      });
-    })
-    .catch(() => {
-      res.status(500).json({ message: "Запрашиваемый ресурс не найден" });
-    });
-};
 
 /*USERS*/
 exports.getAllUsers = (req, res) => {
-  readBase(usersPath)
+  readBase.readBase(usersPath)
     .then((data) => {
       const userSchedule = JSON.parse(data);
       res.status(200).json({
@@ -54,7 +22,7 @@ exports.getAllUsers = (req, res) => {
 };
 /*USERS ID*/
 exports.getUser = (req, res) => {
-  readBase(usersPath)
+  readBase.readBase(usersPath)
     .then((data) => {
       const userSchedule = JSON.parse(data);
       let isThereUser = false;
