+++ bin/server.js
@@ -1,4 +1,4 @@
-const app = require("./app");
+const app = require("../app");
 
 const port = process.env.PORT || 3000;
 app.listen(port);
